+++ CHANGELOG.md
@@ -20,6 +20,8 @@ changes.
 
 - Add API endpoint `POST /transaction` to submit transaction to the head.
 
+- Tested with `cardano-node 10.4.1` and `cardano-cli 10.8.0.0`.
+
 ## [0.22.2] - 2025.06.30
 
 * Fix wrong hydra-script-tx-ids in networks.json